Separate Time Info Parser
